introduce VisualizerUtilities class and move visualization methods inside it.
