+++ Bindings/simulation.i
@@ -231,7 +231,7 @@ OpenSim::ModelComponentSet<OpenSim::Controller>;
 %include <OpenSim/Simulation/StatesTrajectoryReporter.h>
 
 %include <OpenSim/Simulation/SimulationUtilities.h>
-%include <OpenSim/Simulation/DataVisualizer.h>
+%include <OpenSim/Simulation/VisualizerUtilities.h>
 
 // Iterators.
 %template(FrameList) OpenSim::ComponentList<const OpenSim::Frame>;
